rewrite demo3 to use computeTree
